requirements.txt
streamlit
validators
langchain_community
langchain_groq
langchain_huggingface
langchain
faiss-cpu
python-dotenv
